5.14. Модуляция Разделите код, уже написанный в main.js, на отдельные ES2015-модули. main.js станет главным модулем, точкой входа.
