+++ js/data.js
@@ -1,0 +1,37 @@
+const dicts = {
+  descriptions: [
+    'Котики',
+    'Один мой день',
+    'Просто красивая картинка',
+    'На улице',
+    'Красиво?',
+    'Завтрак',
+    'Природа',
+    'Слетал побырому',
+    'На море',
+    'Батонюсь',
+    'Зацените',
+    'Такую вот фиговину увидел',
+    'Без описания'
+  ],
+  messages: [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  ],
+  names: [
+    'Иван',
+    'Хуан Себастьян',
+    'Мария',
+    'Кристоф',
+    'Виктор',
+    'Юлия',
+    'Люпита',
+    'Вашингтон',
+  ],
+};
+
+export default dicts;
